apply FreeBSD ports patches
Taken from: FreeBSD

+++ src/apinger.h
@@ -118,7 +118,7 @@ extern char *config_file;
 
 extern int icmp_sock;
 extern int icmp6_sock;
-extern int ident;
+extern uint16_t ident;
 
 extern struct timeval next_probe;
